Made status bar transparent to better utilize remaining space.
